requirements.txt
Django==1.6.5
Pillow==2.5.1
South==1.0
astroid==1.2.0
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-debug-toolbar==1.2.1
django-extensions==1.3.9
django-redis==3.7.0
djangorestframework==2.3.14
gnureadline==6.3.3
gunicorn==19.1.1
ipdb==0.8
ipython==2.1.0
logilab-common==0.62.1
psycopg2==2.5.3
pylint==1.3.0
pylint-django==0.4
pylint-plugin-utils==0.2.1
redis==2.10.1
requests==2.3.0
sendgrid==0.5.4
six==1.7.3
smtpapi==0.1.2
sqlparse==0.1.11
static3==0.5.1
unittest2==0.5.1
wsgiref==0.1.2
